+++ rust/blackjack/src/strategy/hard.rs
@@ -72,7 +72,7 @@ fn create_hard_total_decisions() -> Vec<HashMap<cards::CardRank, Decision>> {
     for i in 4..22 {
         let mut decisions_row: HashMap<cards::CardRank, Decision> = HashMap::new();
         for rank in cards::CardRank::iterator() {
-            let decision: Decision = _HARD_TOTAL_DECISIONS[i][rank.discriminant()];
+            let decision: Decision = _HARD_TOTAL_DECISIONS[i][rank.discriminant() as usize];
             decisions_row.insert(*rank, decision);
             // the trait `Eq` is not implemented for `CardRank`, which is required by `HashMap<_, _, _>: Index<&_>`
             // the trait `Hash` is not implemented for `CardRank`, which is required by `HashMap<_, _, _>: Index<&_>`
@@ -92,7 +92,7 @@ lazy_static! {
             let mut decisions_row: HashMap<cards::CardRank, Decision> = HashMap::new();
             for rank in cards::CardRank::iterator() {
                 // rank: &cards::CardRank
-                let decision: Decision = _HARD_TOTAL_DECISIONS[i][rank.discriminant()];
+                let decision: Decision = _HARD_TOTAL_DECISIONS[i][rank.discriminant() as usize];
                 decisions_row.insert(*rank, decision);
                 // the trait `Eq` is not implemented for `CardRank`, which is required by `HashMap<_, _, _>: Index<&_>`
                 // the trait `Hash` is not implemented for `CardRank`, which is required by `HashMap<_, _, _>: Index<&_>`
